server/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-editorjs-fields==0.2.7
django-extensions==3.2.3
django-ninja==1.1.0
django-ninja-extra==0.20.5
django-ninja-jwt==5.3.0
gunicorn==21.2.0
injector==0.21.0
mysqlclient==2.2.4
packaging==24.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
sentry-sdk==1.44.1
setuptools==69.2.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
